[rename] feature news 폴더 추가
