+++ src/app/news/page.tsx
@@ -1,5 +1,5 @@
 import { fetchNewsByKeywords } from '@/entities/news'
-import { SearchSection } from '@/features/news-search'
+import { SearchSection } from '@/features/news/news-search'
 import { MajorNews } from '@/widgets/news-major'
 import { NewsCategory } from '@/widgets/news-section'
 import { Metadata } from 'next'
