+++ src/features/news/news-search/ui/SearchInput.tsx
