+++ src/features/news/news-search/ui/SearchSection.tsx
